requirements_dev.txt
annotated-types==0.7.0
anyio==4.10.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
dnspython==2.7.0
email_validator==2.2.0
exceptiongroup==1.3.0
fastapi==0.116.1
flake8==7.3.0
h11==0.16.0
idna==3.10
linecache2==1.0.0
loguru==0.7.3
mccabe==0.7.0
prometheus_client==0.22.1
psycopg2-binary==2.9.10
pycodestyle==2.14.0
pydantic==2.11.7
pydantic_core==2.33.2
pyflakes==3.4.0
python-dotenv==1.1.1
requests==2.32.4
six==1.17.0
sniffio==1.3.1
starlette==0.47.2
traceback2==1.4.0
typing-inspection==0.4.1
typing_extensions==4.14.1
unittest2==1.1.0
urllib3==2.5.0
uvicorn==0.35.0
